# perf: optimize `DataFrame.describe` by presorting columns