This PR works on https://github.com/pola-rs/polars/issues/9368

Presorting columns makes describe much faster especially when the user inputs a lot of percentiles;

The presort will happen once, all quantiles/min/max on numerical columns will be O(1);

Here are some benchmarks on real sales data I am working on:

```
In [18]: df
Out[18]:
shape: (28_061_078, 10)
┌───────────────┬───────────────┬───────────────┬───┬──────────┬──────────┐
│ aaaaa_aaaa_aa ┆ bbbbb_bbbb_bb ┆ ccccc_cccc_cc ┆ … ┆ ddddd_ddd┆ eeee_eee │
│ ---           ┆ ---           ┆ ---           ┆   ┆ ---      ┆ ---      │
│ f64           ┆ f64           ┆ f64           ┆   ┆ f64      ┆ f64      │
╞═══════════════╪═══════════════╪═══════════════╪═══╪══════════╪══════════╡
│ null          ┆ null          ┆ 0.0           ┆ … ┆ 33.38    ┆ 0.0      │
│ null          ┆ null          ┆ 0.409015      ┆ … ┆ 0.0      ┆ 15.55    │
│ null          ┆ null          ┆ 0.0           ┆ … ┆ 18.5     ┆ 0.0      │
│ null          ┆ null          ┆ 0.0           ┆ … ┆ 0.0      ┆ 17.96    │
│ null          ┆ null          ┆ 0.0           ┆ … ┆ 17.36    ┆ 0.0      │
│ …             ┆ …             ┆ …             ┆ … ┆ …        ┆ …        │
│ 77549.54      ┆ 10954.0       ┆ null          ┆ … ┆ null     ┆ null     │
│ null          ┆ null          ┆ null          ┆ … ┆ null     ┆ null     │
│ null          ┆ null          ┆ null          ┆ … ┆ null     ┆ null     │
│ null          ┆ null          ┆ null          ┆ … ┆ null     ┆ null     │
│ 37.44         ┆ 36.0          ┆ null          ┆ … ┆ null     ┆ null     │
└───────────────┴───────────────┴───────────────┴───┴──────────┴──────────┘

In [19]: len(df.columns)
Out[19]: 10

In [12]: %timeit df.describe()
**1.77 s ± 121 ms** per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [13]: %timeit df.describe(percentiles=[0.05, 0.1, 0.2, 0.3, 0.4, 0.5, 0.7, 0.8, 0.9, 0.95, 0.99])
**5.23 s ± 85 ms** per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [14]: %timeit df.select(pl.all().sort()).describe()
**729 ms ± 44.2 ms** per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [15]: %timeit df.select(pl.all().sort()).describe(percentiles=[0.05, 0.1, 0.2, 0.3, 0.4, 0.5, 0.7, 0.8, 0.9, 0.95, 0.99])
**696 ms ± 25.4 ms** per loop (mean ± std. dev. of 7 runs, 1 loop each)
```