@alexander-beedie 2024-01-19
Also: what if the user selects _no_ percentiles? :)