@taki-mekhalfa 2024-01-19
> Thanks! The benchmarks don't really say much without a before/after though - could you do a timing with and without your change? Make sure to use a release build, e.g. `make build-release`

Here is it after a build release native: 
```
In [9]: %timeit df.describe()
1.34 s ± 105 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [10]: %timeit df.describe(percentiles=[0.05, 0.1, 0.2, 0.3, 0.4, 0.5, 0.7, 0.8, 0.9, 0.95, 0.99])
1.26 s ± 57.2 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

Compared to polars `0.20.5`:

```
In [5]: %timeit df.describe()
1.57 s ± 38.6 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [6]: %timeit df.describe(percentiles=[0.05, 0.1, 0.2, 0.3, 0.4, 0.5, 0.7, 0.8, 0.9, 0.95, 0.99])
5.14 s ± 100 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```