@taki-mekhalfa 2024-01-19
> Also: what if the user selects _no_ percentiles? :)
> 
> (This seems likely to be a good idea though; if I recall correctly @orlp was thinking about this during discussions about the potential addition of `n_unique` to the default stats).

```
In [11]: %timeit df.describe(percentiles=[])
743 ms ± 57.2 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

Compared to polars `0.20.5`

```
In [8]: %timeit df.describe(percentiles=[])
267 ms ± 10.9 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

Maybe it's good to not sort if the user did not give any percentiles?