@alexander-beedie 2024-01-19
> Maybe it's good to not sort if the user did not give any percentiles?

That was my suspicion; easy to fix though 😉👍