@orlp 2024-01-19
I think this is sensible - eventually we will be able to get rid of it again once I've gotten to my planned qcut/quantile rework. Then Polars will efficiently be able to support getting multiple quantiles at once. 

Please change it so that it only sorts if more than 1 quantile is requested though.