@taki-mekhalfa 2024-01-19
I made the changes; can you please check? 

```
In [10]: %timeit df.describe(percentiles=[0.05, 0.1, 0.2, 0.3, 0.4, 0.5, 0.7, 0.8, 0.9, 0.95, 0.99])
1.26 s ± 73.4 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [11]: %timeit df.describe(percentiles=[])
268 ms ± 13.2 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```