@alexander-beedie 2024-01-19
> I made the changes; can you please check?

Can you add a `precentiles=[]` to the unit tests?
Looks like that case isn't currently validated.