@taki-mekhalfa 2024-01-19
> > I made the changes; can you please check?
> 
> Can you add a new test that adds coverage for `percentiles=[]`? Looks like that case isn't currently validated.
> 
> See: `py-polars/tests/unit/dataframe/test_describe.py`

There is already a test like that:

```python
@pytest.mark.parametrize("pcts", [None, []])
def test_df_describe_no_percentiles(pcts: list[float] | None) -> None:
```