@alexander-beedie 2024-01-21
> There is already a test like that:

Ahh, well spotted; that would cover it.