@alexander-beedie 2024-01-23
> Is this one good to go @alexander-beedie ?

Yup; I don't mind rebasing my related patch (which we need [you to referee](https://github.com/pola-rs/polars/pull/13720#discussion_r1457834331), lol) on top of it afterwards  ;)